+++ NEWS.md
@@ -2,11 +2,12 @@
 
 # DPI 2025.10
 
-**This version contains breaking changes to function names.**
+**This version contains breaking changes to function names and visualization methods.**
 
--   Added `DPI_dag()`: Directed acyclic graphs (DAGs) via the DPI exploratory analysis for all significant partial correlations.
--   Fixed `cor_network()` to return the correct *p* values of (partial) correlation coefficients.
+-   Added `DPI_dag()`: Directed acyclic graphs (DAGs) via DPI exploratory analysis (causal discovery) for all significant partial correlations.
+-   Added S3 methods `plot.cor.net()`, `plot.bns.dag()`, and `plot.dpi.dag()` that can transform `qgraph` base-plot objects into `ggplot` objects for more stable and flexible visualization.
 -   Renamed `cor_network()` to `cor_net()`, `dag_network()` to `BNs_dag()`, and `matrix_cor()` to `cor_matrix()`.
+-   Fixed `cor_net()` to return the exactly correct *p* values of (partial) correlation coefficients.
 
 # DPI 2025.9